llama-index/llama-alfred-agent/pyproject.toml
[project]
name = "llama-alfred-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "datasets>=4.0.0",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-llms-huggingface-api>=0.6.0",
]
